// web/src/api/generated/model/budgetAlertResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Couplefins
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetAlertResponseHealth } from './budgetAlertResponseHealth';

export interface BudgetAlertResponse {
  group_id: string;
  group_name: string;
  monthly_budget: number;
  monthly_spent: number;
  health: BudgetAlertResponseHealth;
}